pyproject.toml
[tool.poetry]
name = "dream-os"
version = "0.5.0"
description = "Dream.OS - Agentic Development Environment"
authors = ["DreamOS Contributors <maintainers@example.com>"]
readme = "README.md"
packages = [{ include = "dreamos", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9"

# --- Core Libraries ---
PyQt6 = ">=6.7.0"
PyQt6-WebEngine = ">=6.7.0"
PyQt6-sip = ">=13.6.0"
# PyQt5 = "^5.15.9"
# PyQtWebEngine = "^5.15.6"

# Web/Automation
requests = ">=2.28.0"
selenium = ">=4.10.0"
undetected-chromedriver = ">=3.5.0"
webdriver-manager = ">=4.0.0"
beautifulsoup4 = ">=4.12.3"
markdownify = ">=0.11.6"
aiohttp = ">=3.9.5"
playwright = ">=1.40"

# AI/LLM
openai = ">=1.23.6"
jinja2 = ">=3.1.0"

# Data Handling & Validation
pydantic = ">=2.0"
PyYAML = ">=6.0"
pandas = ">=2.0"
jsonschema = ">=4.0.0"

# Async Utilities
async-lru = ">=2.0.4"
APScheduler = ">=3.10,<4.0"

# Web / Visualization
streamlit = ">=1.0"

# CLI Framework
typer = {extras = ["all"], version = ">=0.9.0"}
rich = ">=13.7.1"

# Desktop Interaction (If needed)
pyautogui = ">=0.9.54"
pyperclip = ">=1.8.2"
Pillow = ">=10.3.0"
pywinauto = ">=0.6.8"
pytesseract = "*"

# Misc/Other Core
python-dotenv = ">=0.21.0"
# fastapi = "==0.95.2"
uvicorn = "==0.23.2"
lxml = ">=4.9.0"
praw = ">=7.7.0"
screeninfo = ">=0.7.0"
python-dateutil = "==2.8.2"
pytz = "==2023.3"
numpy = "==1.24.3"
opencv-python = "==4.8.0.74"
loguru = "==0.7.0"
gitpython = "==3.1.31"
azure-storage-blob = ">=12.9.0"
watchdog = ">=2.1.0"
filelock = ">=3.12.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2.0"
pytest-asyncio = ">=0.23.6"
pytest-cov = ">=4.0.0"
ruff = ">=0.1.0"
mypy = ">=1.0.0"
black = ">=24.0.0"
isort = ">=5.12.0"
pre-commit = ">=3.0,<4.0"
flake8 = ">=6.0,<7.0"
flake8-bugbear = "*"
# pytest-qt?

[tool.poetry.scripts]
dream-cli = "dreamos.cli.cli:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]

# Remove the duplicated sections from the previous attempt below this line
# // ... existing code ...

# {{ EDIT START: Add Black and isort config }}
[tool.black]
line-length = 88
target-version = ['py39'] # Align with python dependency

[tool.isort]
profile = "black"
line_length = 88
# {{ EDIT END }}
